MINOR: clean up TestUtils.scala (#17595)
Reviewers: Chia-Ping Tsai <chia7712@gmail.com>
